Switch all time values in X11 to UIntPtr
They're all unsigned at the native layer. This means we should always be treated them as unsigned in the FFI layer

+++ src/Avalonia.X11/X11Window.Ime.cs
@@ -77,7 +77,7 @@ namespace Avalonia.X11
             {
                 (_ime, _imeControl) = ime.Value;
                 _imeControl.Commit += s =>
-                    ScheduleInput(new RawTextInputEventArgs(_keyboard, (ulong)_x11.LastActivityTimestamp.ToInt64(),
+                    ScheduleInput(new RawTextInputEventArgs(_keyboard, _x11.LastActivityTimestamp.ToUInt64(),
                         InputRoot, s));
                 _imeControl.ForwardKey += OnImeControlForwardKey;
             }
@@ -91,7 +91,7 @@ namespace Avalonia.X11
             ScheduleInput(forwardedKey.WithText ?
                 new RawKeyEventArgsWithText(
                     _keyboard,
-                    (ulong)_x11.LastActivityTimestamp.ToInt64(),
+                    _x11.LastActivityTimestamp.ToUInt64(),
                     InputRoot,
                     forwardedKey.Type,
                     X11KeyTransform.KeyFromX11Key(x11Key),
@@ -101,7 +101,7 @@ namespace Avalonia.X11
                     keySymbol) :
                 new RawKeyEventArgs(
                     _keyboard,
-                    (ulong)_x11.LastActivityTimestamp.ToInt64(),
+                    _x11.LastActivityTimestamp.ToUInt64(),
                     InputRoot,
                     forwardedKey.Type,
                     X11KeyTransform.KeyFromX11Key(x11Key),
@@ -117,7 +117,7 @@ namespace Avalonia.X11
             var physicalKey = X11KeyTransform.PhysicalKeyFromScanCode(ev.KeyEvent.keycode);
             var (x11Key, key, symbol) = LookupKey(ref ev.KeyEvent, physicalKey);
             var modifiers = TranslateModifiers(ev.KeyEvent.state);
-            var timestamp = (ulong)ev.KeyEvent.time.ToInt64();
+            var timestamp = ev.KeyEvent.time.ToUInt64();
 
             var args = ev.type == XEventName.KeyPress ?
                 new RawKeyEventArgsWithText(
